pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "faultier"
version = "0.2.2"
#dynamic = ["version"]

authors = [{ name = "Thomas 'stacksmashing' Roth", email = "code@stacksmashing.net"}]
description = "A library to control the Faultier fault-injection tool."
#readme = "README.md"
#license = { file = "LICENSE" }  # Ensure you have a 'LICENSE' file at the root of your project
#classifiers = [
#    "Programming Language :: Python :: 3",
#    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
#    "Operating System :: OS Independent"
#]

dependencies = [
    "pyserial",
    "pyusb",
    "plotly",
    "google",
    "protobuf==5.29.5",
    "tqdm",
    "matplotlib",
    "jupyter",
    # "numpy",
    "dash>=2.14.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
    "dash-bootstrap-components>=1.5.0",
    "numpy>=1.24.0",
    "pyocd-faultier"
]

#[project.urls]
#homepage = ""  # Optionally add project homepage URL
#documentation = ""  # Optionally add documentation URL
#repository = ""  # Optionally add repository URL

[tool.setuptools]
packages = ["faultier"]
include-package-data = false

[tool.setuptools.package-data]
faultier = ["docs/*.svg", "example_firmware/*.hex"]

[project.scripts]
faultier = "faultier:main"
